requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
absl-py==0.10.0           # via tensorboard, tensorflow
appdirs==1.4.4            # via black
asgiref==3.3.0            # via django
astunparse==1.6.3         # via tensorflow
beautifulsoup4==4.9.3     # via pymediawiki
black==20.8b1             # via -r requirements.in
bleach==3.1.5             # via wiki
cachetools==4.1.1         # via google-auth
certifi==2020.6.20        # via requests
chardet==3.0.4            # via requests
click==7.1.2              # via black, sacremoses
cython==0.29.21           # via streamp3
django-classy-tags==2.0.0  # via django-sekizai
django-js-asset==1.2.2    # via django-mptt
django-mptt==0.11.0       # via wiki
django-nyt==1.1.5         # via wiki
django-sekizai==2.0.0     # via wiki
django==3.0.10            # via django-classy-tags, django-mptt, django-nyt, django-sekizai, wiki
filelock==3.0.12          # via transformers
gast==0.3.3               # via tensorflow
google-auth-oauthlib==0.4.1  # via tensorboard
google-auth==1.22.1       # via google-auth-oauthlib, tensorboard
google-pasta==0.2.0       # via tensorflow
grpcio==1.32.0            # via tensorboard, tensorflow
h5py==2.10.0              # via tensorflow
idna==2.10                # via requests
joblib==0.17.0            # via sacremoses
keras-preprocessing==1.1.2  # via tensorflow
markdown==3.2.2           # via tensorboard, wiki
mypy-extensions==0.4.3    # via black
numpy==1.18.5             # via h5py, keras-preprocessing, opt-einsum, spokestack, tensorboard, tensorflow, transformers
oauthlib==3.1.0           # via requests-oauthlib
opt-einsum==3.3.0         # via tensorflow
packaging==20.4           # via bleach, transformers
pathspec==0.8.0           # via black
pillow==8.0.0             # via wiki
protobuf==3.13.0          # via tensorboard, tensorflow, transformers
pyasn1-modules==0.2.8     # via google-auth
pyasn1==0.4.8             # via pyasn1-modules, rsa
pyaudio==0.2.11           # via spokestack
pymediawiki==0.6.6        # via -r requirements.in
pyparsing==2.4.7          # via packaging
pytz==2020.1              # via django
regex==2020.10.15         # via black, sacremoses, transformers
requests-oauthlib==1.3.0  # via google-auth-oauthlib
requests==2.24.0          # via pymediawiki, requests-oauthlib, spokestack, tensorboard, transformers
rsa==4.6                  # via google-auth
sacremoses==0.0.43        # via transformers
sentencepiece==0.1.91     # via transformers
six==1.15.0               # via absl-py, astunparse, bleach, google-auth, google-pasta, grpcio, h5py, keras-preprocessing, packaging, protobuf, sacremoses, tensorboard, tensorflow, websocket-client
sorl-thumbnail==12.6.3    # via wiki
soupsieve==2.0.1          # via beautifulsoup4
spokestack==0.0.7         # via -r requirements.in
sqlparse==0.4.1           # via django
streamp3==0.1.7           # via spokestack
tensorboard-plugin-wit==1.7.0  # via tensorboard
tensorboard==2.3.0        # via tensorflow
tensorflow-estimator==2.3.0  # via tensorflow
tensorflow==2.3.1         # via -r requirements.in
termcolor==1.1.0          # via tensorflow
tokenizers==0.9.2         # via spokestack, transformers
toml==0.10.1              # via black
tqdm==4.50.2              # via sacremoses, transformers
transformers==3.4.0       # via -r requirements.in
typed-ast==1.4.1          # via black
typing-extensions==3.7.4.3  # via black
urllib3==1.25.10          # via requests
webencodings==0.5.1       # via bleach
webrtcvad==2.0.10         # via spokestack
websocket-client==0.57.0  # via spokestack
werkzeug==1.0.1           # via tensorboard
wheel==0.35.1             # via astunparse, tensorboard, tensorflow
wiki==0.6                 # via -r requirements.in
wrapt==1.12.1             # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
